Rename event modules to comply with naming pattern
