+++ .changeset/big-gorillas-fail1.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-events-backend-module-bitbucket-cloud': patch
+---
+
+Renamed `bitbucketCloudEventRouterEventsModule` to `eventsModuleBitbucketCloudEventRouter` to match the [recommended naming patterns](https://backstage.io/docs/backend-system/architecture/naming-patterns).
