+++ .changeset/big-gorillas-fail2.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-events-backend-module-aws-sqs': patch
+---
+
+Renamed `awsSqsConsumingEventPublisherEventsModule` to `eventsModuleAwsSqsConsumingEventPublisher` to match the [recommended naming patterns](https://backstage.io/docs/backend-system/architecture/naming-patterns).
