+++ .changeset/big-gorillas-fail3.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-events-backend-module-gerrit': patch
+---
+
+Renamed `gerritEventRouterEventsModule` to `eventsModuleGerritEventRouter` to match the [recommended naming patterns](https://backstage.io/docs/backend-system/architecture/naming-patterns).
