+++ .changeset/big-gorillas-fail4.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-events-backend-module-github': patch
+---
+
+Renamed `githubEventRouterEventsModule` to `eventsModuleGithubEventRouter` and `githubWebhookEventsModule` to `eventsModuleGithubWebhook`, to match the [recommended naming patterns](https://backstage.io/docs/backend-system/architecture/naming-patterns).
