+++ .changeset/big-gorillas-fail5.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-events-backend-module-gitlab': patch
+---
+
+Renamed `gitlabEventRouterEventsModule` to `eventsModuleGitlabEventRouter` and `gitlabWebhookEventsModule` to `eventsModuleGitlabWebhook` to match the [recommended naming patterns](https://backstage.io/docs/backend-system/architecture/naming-patterns).
