+++ .changeset/big-gorillas-fail6.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-events-backend-module-azure': patch
+---
+
+Renamed `azureDevOpsEventRouterEventsModule` to `eventsModuleAzureDevOpsEventRouter` to match the [recommended naming patterns](https://backstage.io/docs/backend-system/architecture/naming-patterns).
