+++ .changeset/stupid-games-brake.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-events-backend': patch
+---
+
+Updated README instructions
